Move expiring_hash_map tests mod into it's main file
